project/requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==5.0.3
django-database-url==1.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.3
identify==2.5.35
inflection==0.5.1
nodeenv==1.8.0
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
virtualenv==20.25.1
